feat(target_chains/solana): make solana-program optional
