+++ target_chains/solana/programs/pyth-price-publisher/Cargo.toml
@@ -4,8 +4,8 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-bytemuck = "1.13.0"
-solana-program = "=1.14.17"
+bytemuck = { version = "1.13.0", features = ["derive"] }
+solana-program = { version = "=1.14.17", optional = true }
 thiserror = "1.0.40"
 
 [dev-dependencies]
